+++ jaxlib/gpu/solver_kernels_ffi.cc
@@ -18,7 +18,6 @@ limitations under the License.
 #include <algorithm>
 #include <cstdint>
 #include <memory>
-#include <optional>
 #include <string_view>
 
 #if JAX_GPU_HAVE_64_BIT
@@ -51,8 +50,6 @@ namespace JAX_GPU_NAMESPACE {
 
 namespace ffi = ::xla::ffi;
 
-#if JAX_GPU_HAVE_64_BIT
-
 // Map an FFI buffer element type to the appropriate GPU solver type.
 inline absl::StatusOr<gpuDataType> SolverDataType(ffi::DataType dataType,
                                                   std::string_view func) {
@@ -70,8 +67,6 @@ inline absl::StatusOr<gpuDataType> SolverDataType(ffi::DataType dataType,
           "Unsupported dtype %s in %s", absl::FormatStreamed(dataType), func));
   }
 }
-
-#endif
 
 #define SOLVER_DISPATCH_IMPL(impl, ...)           \
   switch (dataType) {                             \
@@ -434,7 +429,8 @@ ffi::Error Syevd64Impl(int64_t batch, int64_t n, gpuStream_t stream,
           params, [](gpusolverDnParams_t p) { gpusolverDnDestroyParams(p); });
 
   int64_t batch_step = 1;
-  FFI_ASSIGN_OR_RETURN(bool is_batched_syev_supported, IsSyevBatchedSupported());
+  FFI_ASSIGN_OR_RETURN(bool is_batched_syev_supported,
+                       IsSyevBatchedSupported());
   if (is_batched_syev_supported) {
     int64_t matrix_size = n * n * ffi::ByteWidth(dataType);
     batch_step = std::numeric_limits<int>::max() / matrix_size;
@@ -450,7 +446,8 @@ ffi::Error Syevd64Impl(int64_t batch, int64_t n, gpuStream_t stream,
         &workspaceInBytesOnHost, std::min(batch, batch_step)));
   } else {
     if (batch_step != 1) {
-      return ffi::Error(ffi::ErrorCode::kInternal,
+      return ffi::Error(
+          ffi::ErrorCode::kInternal,
           "Syevd64Impl: batch_step != 1 but batched syev is not supported");
     }
     JAX_FFI_RETURN_IF_GPU_ERROR(gpusolverDnXsyevd_bufferSize(
@@ -484,17 +481,19 @@ ffi::Error Syevd64Impl(int64_t batch, int64_t n, gpuStream_t stream,
     size_t batch_size = static_cast<size_t>(std::min(batch_step, batch - i));
     if (is_batched_syev_supported) {
       JAX_FFI_RETURN_IF_GPU_ERROR(gpusolverDnXsyevBatched(
-          handle.get(), params, jobz, uplo, n, aType, out_data, n, wType, w_data,
-          aType, workspaceOnDevice, workspaceInBytesOnDevice,
-          workspaceOnHost.get(), workspaceInBytesOnHost, info_data, batch_size));
+          handle.get(), params, jobz, uplo, n, aType, out_data, n, wType,
+          w_data, aType, workspaceOnDevice, workspaceInBytesOnDevice,
+          workspaceOnHost.get(), workspaceInBytesOnHost, info_data,
+          batch_size));
     } else {
       if (batch_step != 1) {
-        return ffi::Error(ffi::ErrorCode::kInternal,
+        return ffi::Error(
+            ffi::ErrorCode::kInternal,
             "Syevd64Impl: batch_step != 1 but batched syev is not supported");
       }
       JAX_FFI_RETURN_IF_GPU_ERROR(gpusolverDnXsyevd(
-          handle.get(), params, jobz, uplo, n, aType, out_data, n, wType, w_data,
-          aType, workspaceOnDevice, workspaceInBytesOnDevice,
+          handle.get(), params, jobz, uplo, n, aType, out_data, n, wType,
+          w_data, aType, workspaceOnDevice, workspaceInBytesOnDevice,
           workspaceOnHost.get(), workspaceInBytesOnHost, info_data));
     }
     out_data += out_step;
@@ -1214,6 +1213,132 @@ XLA_FFI_DEFINE_HANDLER_SYMBOL(SytrdFfi, SytrdDispatch,
                                   .Ret<ffi::AnyBuffer>()         // tau
                                   .Ret<ffi::Buffer<ffi::S32>>()  // info
 );
+
+// General eigenvalue decomposition: geev
+
+#if JAX_GPU_HAVE_SOLVER_GEEV
+
+ffi::Error GeevImpl(gpuStream_t stream, ffi::ScratchAllocator scratch,
+                    bool left, bool right, ffi::AnyBuffer a,
+                    ffi::Result<ffi::AnyBuffer> out,
+                    ffi::Result<ffi::AnyBuffer> w,
+                    ffi::Result<ffi::AnyBuffer> vl,
+                    ffi::Result<ffi::AnyBuffer> vr,
+                    ffi::Result<ffi::Buffer<ffi::S32>> info) {
+  auto dataType = a.element_type();
+  if (dataType != vr->element_type()) {
+    return ffi::Error::InvalidArgument(
+        "The inputs and outputs to geev must have the same element type");
+  }
+  if (dataType != w->element_type() &&
+      !(dataType == ffi::F32 && w->element_type() == ffi::C64) &&
+      !(dataType == ffi::F64 && w->element_type() == ffi::C128)) {
+    return ffi::Error::InvalidArgument(
+        "The eigenvector output type of geev must match the input type or "
+        "be its complex counterpart.");
+  }
+
+  FFI_ASSIGN_OR_RETURN((auto [batch, m, n]), SplitBatch2D(a.dimensions()));
+  if (m != n) {
+    return ffi::Error::InvalidArgument(
+        "The input matrix to geev must be square");
+  }
+  int w_len;
+  if (w->element_type() == ffi::F32 || w->element_type() == ffi::F64) {
+    w_len = 2 * n;
+    FFI_RETURN_IF_ERROR(
+        CheckShape(w->dimensions(), {batch, 2 * n}, "w", "geev"));
+  } else {
+    FFI_RETURN_IF_ERROR(CheckShape(w->dimensions(), {batch, n}, "w", "geev"));
+    w_len = n;
+  }
+  if (left) {
+    FFI_RETURN_IF_ERROR(
+        CheckShape(vl->dimensions(), {batch, n, n}, "vl", "geev"));
+  }
+  if (right) {
+    FFI_RETURN_IF_ERROR(
+        CheckShape(vr->dimensions(), {batch, n, n}, "vr", "geev"));
+  }
+  FFI_RETURN_IF_ERROR(CheckShape(info->dimensions(), batch, "info", "geev"));
+
+  FFI_ASSIGN_OR_RETURN(auto handle, SolverHandlePool::Borrow(stream));
+  FFI_ASSIGN_OR_RETURN(auto aType, SolverDataType(dataType, "geev"));
+  FFI_ASSIGN_OR_RETURN(auto wType, SolverDataType(w->element_type(), "geev"));
+
+  // At the time of writing, cusolver only supports computing right
+  // eigenvectors, but has the option for left eigenvectors in its API. Let us
+  // assume that they intend to add support for left eigenvectors in the future.
+  gpusolverEigMode_t jobvl =
+      left ? GPUSOLVER_EIG_MODE_VECTOR : GPUSOLVER_EIG_MODE_NOVECTOR;
+  gpusolverEigMode_t jobvr =
+      right ? GPUSOLVER_EIG_MODE_VECTOR : GPUSOLVER_EIG_MODE_NOVECTOR;
+
+  gpusolverDnParams_t params;
+  JAX_FFI_RETURN_IF_GPU_ERROR(gpusolverDnCreateParams(&params));
+  std::unique_ptr<gpusolverDnParams, void (*)(gpusolverDnParams_t)>
+      params_cleanup(
+          params, [](gpusolverDnParams_t p) { gpusolverDnDestroyParams(p); });
+
+  size_t workspaceInBytesOnDevice, workspaceInBytesOnHost;
+  JAX_FFI_RETURN_IF_GPU_ERROR(cusolverDnXgeev_bufferSize(
+      handle.get(), params, jobvl, jobvr, n, aType, /*a=*/nullptr, n, wType,
+      /*w=*/nullptr, aType, /*vl=*/nullptr, n, aType, /*vr=*/nullptr, n, aType,
+      &workspaceInBytesOnDevice, &workspaceInBytesOnHost));
+
+  auto maybe_workspace = scratch.Allocate(workspaceInBytesOnDevice);
+  if (!maybe_workspace.has_value()) {
+    return ffi::Error(ffi::ErrorCode::kResourceExhausted,
+                      "Unable to allocate device workspace for syevd");
+  }
+  auto workspaceOnDevice = maybe_workspace.value();
+  auto workspaceOnHost =
+      std::unique_ptr<char[]>(new char[workspaceInBytesOnHost]);
+
+  const char* a_data = static_cast<const char*>(a.untyped_data());
+  char* out_data = static_cast<char*>(out->untyped_data());
+  char* w_data = static_cast<char*>(w->untyped_data());
+  char* vl_data = static_cast<char*>(vl->untyped_data());
+  char* vr_data = static_cast<char*>(vr->untyped_data());
+  int* info_data = info->typed_data();
+  if (a_data != out_data) {
+    JAX_FFI_RETURN_IF_GPU_ERROR(gpuMemcpyAsync(
+        out_data, a_data, a.size_bytes(), gpuMemcpyDeviceToDevice, stream));
+  }
+
+  size_t out_step = n * n * ffi::ByteWidth(dataType);
+  size_t w_step = w_len * ffi::ByteWidth(w->element_type());
+
+  for (auto i = 0; i < batch; ++i) {
+    JAX_FFI_RETURN_IF_GPU_ERROR(gpusolverDnXgeev(
+        handle.get(), params, jobvl, jobvr, n, aType, out_data, n, wType,
+        w_data, aType, vl_data, n, aType, vr_data, n, aType, workspaceOnDevice,
+        workspaceInBytesOnDevice, workspaceOnHost.get(), workspaceInBytesOnHost,
+        info_data));
+    out_data += out_step;
+    w_data += w_step;
+    vr_data += out_step;
+    ++info_data;
+  }
+
+  return ffi::Error::Success();
+}
+
+XLA_FFI_DEFINE_HANDLER_SYMBOL(GeevFfi, GeevImpl,
+                              ffi::Ffi::Bind()
+                                  .Ctx<ffi::PlatformStream<gpuStream_t>>()
+                                  .Ctx<ffi::ScratchAllocator>()
+                                  .Attr<bool>("left")
+                                  .Attr<bool>("right")
+                                  .Arg<ffi::AnyBuffer>()         // a
+                                  .Ret<ffi::AnyBuffer>()         // out
+                                  .Ret<ffi::AnyBuffer>()         // w
+                                  .Ret<ffi::AnyBuffer>()         // vl
+                                  .Ret<ffi::AnyBuffer>()         // vr
+                                  .Ret<ffi::Buffer<ffi::S32>>()  // info
+);
+
+#endif  // JAX_GPU_HAVE_SOLVER_GEEV
 
 #undef SOLVER_DISPATCH_IMPL
 #undef SOLVER_BLAS_DISPATCH_IMPL